requirements.txt
akismet==1.1
anyascii==0.2.0
asgiref==3.4.1
attrs==21.2.0
autopep8==1.5.7
beautifulsoup4==4.9.3
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.3
colorclass==2.2.0
coverage==5.5
cryptography==2.8
defusedxml==0.7.1
Django==3.2.5
django-allauth==0.45.0
django-background-tasks==1.2.5
django-classy-tags==2.0.0
django-clearcache==1.2.1
django-compat==1.0.15
django-debug-toolbar==3.2.2
django-environ==0.8.1
django-filter==2.4.0
django-haystack==3.0
django-modelcluster==5.1
django-mysql==3.12.0
django-query-profiler==0.8
django-ratelimit==3.0.1
django-sekizai==2.0.0
django-silk==4.1.0
django-taggit==1.5.1
django-treebeard==4.5.1
django-watson==1.6.2
djangorestframework==3.12.4
dnspython==2.1.0
docopt==0.6.2
draftjs-exporter==2.1.7
elasticsearch==5.5.3
email-validator==1.1.3
et-xmlfile==1.1.0
fuzzywuzzy==0.18.0
gprof2dot==2021.2.21
html5lib==1.1
httplib2==0.20.4
idna==3.2
importlib-metadata==4.6.1
iniconfig==1.1.1
install==1.3.4
isort==5.9.3
jdcal==1.4.1
Jinja2==3.0.1
l18n==2020.6.1
MarkupSafe==2.0.1
mmh3==3.0.0
mo-future==5.17.21182
model-bakery==1.3.2
more-itertools==8.8.0
moz-sql-parser==3.32.20026
mysql-connector-python==8.0.26
mysqlclient==2.0.3
oauthlib==3.1.1
openpyxl==3.0.7
packaging==21.0
Pillow==8.3.1
pip-upgrader==1.4.15
pluggy==0.13.1
pprintpp==0.4.0
protobuf==3.17.3
psutil==5.8.0
py==1.10.0
pycodestyle==2.7.0
pycountry==20.7.3
pycountry-convert==0.7.2
pycparser==2.20
Pygments==2.9.0
PyJWT==2.1.0
pylibmc==1.6.1
pyparsing==3.0.7
pytest==6.2.4
pytest-cov==2.12.1
pytest-mock==3.6.1
python-binary-memcached==0.30.1
python-dateutil==2.8.2
python-Levenshtein==0.12.2
python3-openid==3.2.0
pytz==2021.1
PyYAML==5.4.1
redis==3.5.3
repoze.lru==0.7
requests==2.26.0
requests-oauthlib==1.3.0
shortuuid==1.0.8
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
stripe==2.71.0
tablib==3.0.0
telepath==0.2
terminaltables==3.1.0
toml==0.10.2
ua-parser==0.10.0
uhashring==2.1
Unidecode==1.2.0
urllib3==1.26.6
user-agents==2.2.0
wagtail==2.13.4
wcwidth==0.2.5
webencodings==0.5.1
Whoosh==2.7.4
Willow==1.4
xlrd==2.0.1
XlsxWriter==1.4.4
xlwt==1.3.0
zipp==3.5.0
